refactor user specs
