+++ spec/models/user_part_app_spec.rb
@@ -1,55 +1,34 @@
-#require 'ostruct'
 require_relative '../spec_helper'
-#require_relative 'user_shared_examples'
-#require_relative '../../services/dataservices-metrics/lib/isolines_usage_metrics'
-#require_relative '../../services/dataservices-metrics/lib/observatory_snapshot_usage_metrics'
-#require_relative '../../services/dataservices-metrics/lib/observatory_general_usage_metrics'
-#require 'factories/organizations_contexts'
-#require_relative '../../app/model_factories/layer_factory'
-#require_dependency 'cartodb/redis_vizjson_cache'
-#require 'helpers/rate_limits_helper'
-#require 'helpers/unique_names_helper'
-#require 'helpers/account_types_helper'
-#require 'factories/users_helper'
-#require 'factories/database_configuration_contexts'
-
+require_relative 'user_shared_examples'
+require 'helpers/user_part_helper'
 
 describe User do
-  #include UniqueNamesHelper
-  #include AccountTypesHelper
-  #include RateLimitsHelper
+  include UserPartHelper
+  include_context 'user spec configuration'
 
-  before(:each) do
-    CartoDB::UserModule::DBService.any_instance.stubs(:enable_remote_db_user).returns(true)
+  it "should only allow legal usernames" do
+    illegal_usernames = %w(si$mon 'sergio estella' j@vi sergio£££ simon_tokumine SIMON Simon jose.rilla -rilla rilla-)
+    legal_usernames   = %w(simon javier-de-la-torre sergio-leiva sergio99)
+
+    illegal_usernames.each do |name|
+      @user.username = name
+      @user.valid?.should be_false
+      @user.errors[:username].should be_present
+    end
+
+    legal_usernames.each do |name|
+      @user.username = name
+      @user.valid?.should be_true
+      @user.errors[:username].should be_blank
+    end
   end
 
-  before(:all) do
-    bypass_named_maps
-
-    @user_password = 'admin123'
-    puts "\n[rspec][user_spec] Creating test user databases..."
-    @user     = create_user :email => 'admin@example.com', :username => 'admin', :password => @user_password
-    @user2    = create_user :email => 'user@example.com',  :username => 'user',  :password => 'user123'
-
-    puts "[rspec][user_spec] Loading user data..."
-    reload_user_data(@user) && @user.reload
-
-    puts "[rspec][user_spec] Running..."
-  end
-
-  before(:each) do
-    bypass_named_maps
-    CartoDB::Varnish.any_instance.stubs(:send_command).returns(true)
-    CartoDB::UserModule::DBService.any_instance.stubs(:enable_remote_db_user).returns(true)
-    Table.any_instance.stubs(:update_cdb_tablemetadata)
-  end
-
-  after(:all) do
-    bypass_named_maps
-    @user.destroy
-    @user2.destroy
-    @account_type.destroy if @account_type
-    @account_type_org.destroy if @account_type_org
+  it "should not allow a username in use by an organization" do
+    org = create_org('testusername', 10.megabytes, 1)
+    @user.username = org.name
+    @user.valid?.should be_false
+    @user.username = 'wadus'
+    @user.valid?.should be_true
   end
 
   it "should have a default dashboard_viewed? false" do
@@ -326,86 +305,22 @@ describe User do
   end
 
   describe 'refactored behaviour' do
-  it_behaves_like 'user models' do
-    def get_twitter_imports_count_by_user_id(user_id)
-      get_user_by_id(user_id).get_twitter_imports_count
+    it_behaves_like 'user models' do
+      def get_twitter_imports_count_by_user_id(user_id)
+        get_user_by_id(user_id).get_twitter_imports_count
+      end
+
+      def get_user_by_id(user_id)
+        ::User.where(id: user_id).first
+      end
+
+      def create_user
+        FactoryGirl.create(:valid_user)
+      end
     end
-
-    def get_user_by_id(user_id)
-      ::User.where(id: user_id).first
-    end
-
-    def create_user
-      FactoryGirl.create(:valid_user)
-    end
-  end
-end
-
-describe User do
-  include UniqueNamesHelper
-  include AccountTypesHelper
-  include RateLimitsHelper
-
-  before(:each) do
-    CartoDB::UserModule::DBService.any_instance.stubs(:enable_remote_db_user).returns(true)
-  end
-
-  before(:all) do
-    bypass_named_maps
-
-    @user_password = 'admin123'
-    puts "\n[rspec][user_spec] Creating test user databases..."
-    @user     = create_user :email => 'admin@example.com', :username => 'admin', :password => @user_password
-    @user2    = create_user :email => 'user@example.com',  :username => 'user',  :password => 'user123'
-
-    puts "[rspec][user_spec] Loading user data..."
-    reload_user_data(@user) && @user.reload
-
-    puts "[rspec][user_spec] Running..."
-  end
-
-  before(:each) do
-    bypass_named_maps
-    CartoDB::Varnish.any_instance.stubs(:send_command).returns(true)
-    CartoDB::UserModule::DBService.any_instance.stubs(:enable_remote_db_user).returns(true)
-    Table.any_instance.stubs(:update_cdb_tablemetadata)
-  end
-
-  after(:all) do
-    bypass_named_maps
-    @user.destroy
-    @user2.destroy
-    @account_type.destroy if @account_type
-    @account_type_org.destroy if @account_type_org
-  end
-
-  it "should only allow legal usernames" do
-    illegal_usernames = %w(si$mon 'sergio estella' j@vi sergio£££ simon_tokumine SIMON Simon jose.rilla -rilla rilla-)
-    legal_usernames   = %w(simon javier-de-la-torre sergio-leiva sergio99)
-
-    illegal_usernames.each do |name|
-      @user.username = name
-      @user.valid?.should be_false
-      @user.errors[:username].should be_present
-    end
-
-    legal_usernames.each do |name|
-      @user.username = name
-      @user.valid?.should be_true
-      @user.errors[:username].should be_blank
-    end
-  end
-
-  it "should not allow a username in use by an organization" do
-    org = create_org('testusername', 10.megabytes, 1)
-    @user.username = org.name
-    @user.valid?.should be_false
-    @user.username = 'wadus'
-    @user.valid?.should be_true
   end
 
   describe 'session' do
-
     before(:all) do
       Cartodb::Central.stubs(:sync_data_with_cartodb_central?).returns(true)
       @user = FactoryGirl.create(:valid_user)
@@ -449,25 +364,5 @@ describe User do
         @user.invalidate_all_sessions!
       end
     end
-  end
-
-
-  protected
-
-  def create_org(org_name, org_quota, org_seats)
-    organization = Organization.new
-    organization.name = unique_name(org_name)
-    organization.quota_in_bytes = org_quota
-    organization.seats = org_seats
-    organization.save
-    organization
-  end
-
-  def tables_including_shared(user)
-    Carto::VisualizationQueryBuilder
-      .new
-      .with_owned_by_or_shared_with_user_id(user.id)
-      .with_type(Carto::Visualization::TYPE_CANONICAL)
-      .build.map(&:table)
   end
 end
