+++ spec/models/user_part_organization_spec.rb
@@ -1,55 +1,9 @@
-#require 'ostruct'
 require_relative '../spec_helper'
-#require_relative 'user_shared_examples'
-#require_relative '../../services/dataservices-metrics/lib/isolines_usage_metrics'
-#require_relative '../../services/dataservices-metrics/lib/observatory_snapshot_usage_metrics'
-#require_relative '../../services/dataservices-metrics/lib/observatory_general_usage_metrics'
-#require 'factories/organizations_contexts'
-#require_relative '../../app/model_factories/layer_factory'
-#require_dependency 'cartodb/redis_vizjson_cache'
-#require 'helpers/rate_limits_helper'
-#require 'helpers/unique_names_helper'
-#require 'helpers/account_types_helper'
-#require 'factories/users_helper'
-#require 'factories/database_configuration_contexts'
+require 'helpers/user_part_helper'
 
 describe User do
-  #include UniqueNamesHelper
-  #include AccountTypesHelper
-  #include RateLimitsHelper
-
-  before(:each) do
-    CartoDB::UserModule::DBService.any_instance.stubs(:enable_remote_db_user).returns(true)
-  end
-
-  before(:all) do
-    bypass_named_maps
-
-    @user_password = 'admin123'
-    puts "\n[rspec][user_spec] Creating test user databases..."
-    @user     = create_user :email => 'admin@example.com', :username => 'admin', :password => @user_password
-    @user2    = create_user :email => 'user@example.com',  :username => 'user',  :password => 'user123'
-
-    puts "[rspec][user_spec] Loading user data..."
-    reload_user_data(@user) && @user.reload
-
-    puts "[rspec][user_spec] Running..."
-  end
-
-  before(:each) do
-    bypass_named_maps
-    CartoDB::Varnish.any_instance.stubs(:send_command).returns(true)
-    CartoDB::UserModule::DBService.any_instance.stubs(:enable_remote_db_user).returns(true)
-    Table.any_instance.stubs(:update_cdb_tablemetadata)
-  end
-
-  after(:all) do
-    bypass_named_maps
-    @user.destroy
-    @user2.destroy
-    @account_type.destroy if @account_type
-    @account_type_org.destroy if @account_type_org
-  end
+  include UserPartHelper
+  include_context 'user spec configuration'
 
   describe 'organization checks' do
     it "should not be valid if their organization doesn't have more seats" do
@@ -418,16 +372,4 @@ describe User do
       org.destroy
     end
   end
-
-  protected
-
-  def create_org(org_name, org_quota, org_seats)
-    organization = Organization.new
-    organization.name = unique_name(org_name)
-    organization.quota_in_bytes = org_quota
-    organization.seats = org_seats
-    organization.save
-    organization
-  end
-
 end
